Extract generations methods in fake aggregator build script to a dedicated crate
