+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
   "examples/client-mithril-stake-distribution",
   "examples/client-snapshot",
   "mithril-aggregator",
+  "mithril-build-script",
   "mithril-client",
   "mithril-client-cli",
   "mithril-client-wasm",
