+++ mithril-build-script/.gitignore
@@ -1,0 +1,3 @@
+target/
+.DS_Store
+
