+++ mithril-build-script/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "mithril-build-script"
+version = "0.1.0"
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
