+++ mithril-test-lab/mithril-aggregator-fake/Cargo.toml
@@ -33,4 +33,4 @@ reqwest = "0.11.23"
 warp = "0.3.6"
 
 [build-dependencies]
-serde_json = "1.0.113"
+mithril-build-script = { path = "../../mithril-build-script" }
